+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v2.5.0
+
+* added PHP 8.4 and 8.5 support
+* fixed textarea line-up/down bug when cursor is on start of line
+* updated build tools for linting
+* migrated paginate macro from theme to core
+* fixed SASS @import deprecations
+
 # v2.4.0
 
 * added dark mode for elegant theme